+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.3.0](https://github.com/ErezNagar/lease-calculator/compare/v1.2.2...v1.3.0) - 2/1/2021
+
+### Added
+
+- Support for zero drive-off
+
+### Updated
+
+- Acquisition fee is no longer capitalized by default
+- Acquisition fee and disposition fee are included in drive-off payment when taxMethod = TaxationMethod.TAX_ON_TOTAL_LEASE_PAYMENT
+
 ## [1.2.1](https://github.com/ErezNagar/lease-calculator/compare/v1.2.0...v1.2.1) - 8/11/2020
 
 ### Added